2022.10.13 - 2 (database ready)

+++ app/src/main/kotlin/ir/mahdiparastesh/homechat/data/Message.kt
@@ -13,7 +13,7 @@ import androidx.room.PrimaryKey
     )]
 )
 class Message(
-    val chat: Long,
+    val chat: Short,
     val type: Byte,
 ) {
     @PrimaryKey(autoGenerate = true)
